Add bat theme; upgrade fzf opts
As the title says, this commit adds Catppuccin Macchiato as a theme for
bat to match everything else. This should ensure that the color scheme
is consistent across all tools. It also does the same for fzf, as well
as adding some smarter defaults.

This commit also removes a bit of kruft from the zshrc file.